+++ docs/root/intro/version_history.rst
@@ -59,8 +59,8 @@ Version history
 * router: added :ref:`rq_retry_skipped_request_not_complete <config_http_filters_router_stats>` counter stat to router stats.
 * router: :ref:`Scoped routing <arch_overview_http_routing_route_scope>` is supported.
 * router: added new :ref:`retriable-headers <config_http_filters_router_x-envoy-retry-on>` retry policy. Retries can now be configured to trigger by arbitrary response header matching.
-* router: added ability to reverse header evaluation order, see :ref:`route configuration's reverse header
-  evaluation order flag <envoy_api_field_RouteConfiguration.eval_most_specific_header_mutations_first>`
+* router: added ability for most specific header mutations to take precedence, see :ref:`route configuration's most specific
+  header mutations wins flag <envoy_api_field_RouteConfiguration.most_specific_header_mutations_wins>`
 * router: added new :ref:`retriable request headers <envoy_api_field_route.RetryPolicy.retriable_request_headers>` to retry policies. Retries can now be configured to only trigger on request header match.
 * router check tool: add coverage reporting & enforcement.
 * router check tool: add comprehensive coverage reporting.